https://github.com/mP1/walkingkooka-spreadsheet/pull/3515 moved package reference.store to store
- https://github.com/mP1/walkingkooka-spreadsheet/pull/3515
- moved package reference.store to store
