+++ src/main/java/walkingkooka/spreadsheet/server/label/http/SpreadsheetLabelHateosHandlerOne.java
@@ -20,7 +20,7 @@ package walkingkooka.spreadsheet.server.label.http;
 import walkingkooka.net.http.server.HttpRequestAttribute;
 import walkingkooka.net.http.server.hateos.HateosHandler;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelMapping;
-import walkingkooka.spreadsheet.reference.store.SpreadsheetLabelStore;
+import walkingkooka.spreadsheet.store.SpreadsheetLabelStore;
 
 import java.util.Map;
 import java.util.Optional;
