+++ src/main/java/walkingkooka/spreadsheet/server/label/http/SpreadsheetLabelHateosHandlers.java
@@ -21,7 +21,7 @@ import walkingkooka.net.http.server.hateos.HateosHandler;
 import walkingkooka.reflect.PublicStaticHelper;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelMapping;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelName;
-import walkingkooka.spreadsheet.reference.store.SpreadsheetLabelStore;
+import walkingkooka.spreadsheet.store.SpreadsheetLabelStore;
 
 public final class SpreadsheetLabelHateosHandlers implements PublicStaticHelper {
